+++ src/zopfli/gzip_container.c
@@ -56,7 +56,7 @@ the updated crc. The crc should be initialized to zero.
 static unsigned long UpdateCRC(unsigned long crc,
                                const unsigned char *buf, size_t len) {
   unsigned long c = crc ^ 0xffffffffL;
-  unsigned n;
+  size_t n;
 
   if (!crc_table_computed)
     MakeCRCTable();
@@ -67,7 +67,7 @@ static unsigned long UpdateCRC(unsigned long crc,
 }
 
 /* Returns the CRC of the bytes buf[0..len-1]. */
-static unsigned long CRC(const unsigned char* buf, int len) {
+static unsigned long CRC(const unsigned char* buf, size_t len) {
   return UpdateCRC(0L, buf, len);
 }
